BrownieAtelierStorage/requirements.txt
adal
autopep8
azure-common
azure-core
azure-functions
azure-identity
azure-keyvault-secrets
azure-mgmt-containerinstance
azure-mgmt-core
azure-mgmt-resource
azure-storage-blob
azure-storage-file-share
azure-storage-queue
certifi
cffi
charset-normalizer
cryptography
dnspython
idna
isodate
msal
msal-extensions
msrest
msrestazure
oauthlib
portalocker
pycodestyle
pycparser
pydantic
PyJWT
python-dateutil
python-decouple
requests
requests-oauthlib
six
tomli
typing-extensions
urllib3
